seed et algo

+++ db/migrate/20220831100026_remove_fielexperiences_from_users.rb
@@ -1,16 +1,16 @@
 class RemoveFielexperiencesFromUsers < ActiveRecord::Migration[7.0]
   def change
-    remove_column :users, :vineux, :integer, default: 0
-    remove_column :users, :epicee, :integer, default: 0
-    remove_column :users, :boise, :integer, default: 0
-    remove_column :users, :animale, :integer, default: 0
-    remove_column :users, :noix, :integer, default: 0
-    remove_column :users, :sucre, :integer, default: 0
-    remove_column :users, :fruite, :integer, default: 0
-    remove_column :users, :floral, :integer, default: 0
-    remove_column :users, :herbace, :integer, default: 0
-    remove_column :users, :cereale, :integer, default: 0
-    remove_column :users, :empyreumatique, :integer, default: 0
-    remove_column :users, :tourbe, :integer, default: 0
+    remove_column :users, :vineux, :float, default: 0
+    remove_column :users, :epicee, :float, default: 0
+    remove_column :users, :boise, :float, default: 0
+    remove_column :users, :animale, :float, default: 0
+    remove_column :users, :noix, :float, default: 0
+    remove_column :users, :sucre, :float, default: 0
+    remove_column :users, :fruite, :float, default: 0
+    remove_column :users, :floral, :float, default: 0
+    remove_column :users, :herbace, :float, default: 0
+    remove_column :users, :cereale, :float, default: 0
+    remove_column :users, :empyreumatique, :float, default: 0
+    remove_column :users, :tourbe, :float, default: 0
   end
 end
